# Add Rust for Linux `auto` CI job