@rust-log-analyzer 2024-05-17
The job **`rfl`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/9128072918/job/25099702623) [(plain)](https://github.com/rust-lang/rust/commit/b9d6e9a5eab32aafcaf656ba6b0d3a579acedc4b/checks/25099702623/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
Getting action download info
Download action repository 'msys2/setup-msys2@v2.22.0' (SHA:cc11e9188b693c2b100158c3322424c4cc1dadea)
Download action repository 'actions/checkout@v4' (SHA:0ad4b8fadaa221de15dcec353f45205ec38ea70b)
Download action repository 'actions/upload-artifact@v4' (SHA:65462800fd760344b1a7b4382951275a0abb4808)
Complete job name: PR - rfl
git config --global core.autocrlf false
shell: /usr/bin/bash --noprofile --norc -e -o pipefail {0}
---
      --llvm-root=/usr/lib/llvm-17 \
      --enable-llvm-link-shared \
      --set rust.thin-lto-import-instr-limit=10

COPY /scripts/rfl-build.sh /tmp/rfl-build.sh
ENV SCRIPT bash /tmp/rfl-build.sh

set -euo pipefail

LINUX_VERSION=v6.9
LINUX_VERSION=v6.9

# Build rustc
../x.py --stage 1 build library/std
BUILT_RUSTC=$(realpath ./build/x86_64-unknown-linux-gnu/stage1/bin/rustc)
# Install rustup so that we have cargo for installing bindgen
# Install rustup so that we have cargo for installing bindgen
curl --proto '=https' --tlsv1.2 -sSf -o rustup.sh https://sh.rustup.rs
sh rustup.sh -y --default-toolchain stable --profile minimal
# Install bindgen
# Install bindgen
"$HOME"/.cargo/bin/cargo install --locked --version $(/rfl/linux/scripts/min-tool-version.sh bindgen) bindgen-cli
mkdir -p /rfl
cd /rfl

# Download Linux
# Download Linux
git clone --depth 1 --branch ${LINUX_VERSION} https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
# Configure Linux
# Configure Linux
cat <<EOF > linux/kernel/configs/rfl-for-rust-ci.config
# CONFIG_RETHUNK is not set
# CONFIG_X86_KERNEL_IBT is not set

CONFIG_RUST=y

CONFIG_SAMPLES=y
---
CONFIG_KUNIT=y
CONFIG_RUST_KERNEL_DOCTESTS=y
EOF

export RUSTC=${BUILT_RUSTC}

make -C linux LLVM=1 -j$(($(nproc) + 1)) rustavailable defconfig rfl-for-rust-ci.config
make -C linux LLVM=1 -j$(($(nproc) + 1)) samples/rust/rust_minimal.o samples/rust/rust_print.o drivers/net/phy/ax88796b_rust.o

# ignore-tidy-linelength

set -ex
---
Image input checksum 04bda925e701ede911f878678e7d1abcde08540c19450e3fd747538eb8f0016276c7008efcf6c9996129f735903633eb932c4331ba4b9e7c9f7fff7aa1de8648
##[group]Building docker image for rfl
Docker version 24.0.9, build 2936816
quirky_allen
Attempting with retry: docker buildx build --rm -t rust-ci -f /home/runner/work/rust/rust/src/ci/docker/host-x86_64/rfl/Dockerfile --cache-from type=registry,ref=ghcr.io/rust-lang-ci/rust-ci-cache:04bda925e701ede911f878678e7d1abcde08540c19450e3fd747538eb8f0016276c7008efcf6c9996129f735903633eb932c4331ba4b9e7c9f7fff7aa1de8648 --output=type=docker /home/runner/work/rust/rust/src/ci/docker

#1 [internal] booting buildkit
#1 pulling image moby/buildkit:buildx-stable-1
#1 pulling image moby/buildkit:buildx-stable-1 0.2s done
---
100 22.0M  100 22.0M    0     0   149M      0 --:--:-- --:--:-- --:--:--  149M
#11 0.184 + chmod +x /usr/local/bin/sccache
#11 DONE 0.2s

#12 [5/5] COPY /scripts/rfl-build.sh /tmp/rfl-build.sh

#13 exporting to docker image format
#13 exporting layers
#13 exporting layers 39.5s done
---
DirectMap4k:      192448 kB
DirectMap2M:     8196096 kB
DirectMap1G:    27262976 kB
##[endgroup]
Executing bash /tmp/rfl-build.sh
+ bash /tmp/rfl-build.sh
    Finished `dev` profile [unoptimized] target(s) in 0.03s
##[endgroup]
##[group]Building stage0 library artifacts (x86_64-unknown-linux-gnu)
    Updating crates.io index
---
    Finished `release` profile [optimized] target(s) in 57.26s
##[endgroup]
Build completed successfully in 0:06:11
info: downloading installer
error: $HOME differs from euid-obtained home directory: you may be using sudo
error: $HOME directory: /home/user
error: euid-obtained home directory: /home/ubuntu
info: profile set to 'minimal'
info: default host triple is x86_64-unknown-linux-gnu
info: syncing channel updates for 'stable-x86_64-unknown-linux-gnu'
info: latest update on 2024-05-02, rust version 1.78.0 (9b00956e5 2024-04-29)
info: downloading component 'rust-std'
info: downloading component 'rustc'
info: installing component 'cargo'
info: installing component 'rust-std'
info: installing component 'rust-std'
info: installing component 'rustc'

info: default toolchain set to 'stable-x86_64-unknown-linux-gnu'
  stable-x86_64-unknown-linux-gnu installed - rustc 1.78.0 (9b00956e5 2024-04-29)


Rust is installed now. Great!
To get started you may need to restart your current shell.
This would reload your PATH environment variable to include
This would reload your PATH environment variable to include
Cargo's bin directory (/cargo/bin).
To configure your current shell, you need to source
the corresponding env file under /cargo.


This is usually done by running one of the following (note the leading DOT):
. "/cargo/env"            # For sh/bash/zsh/ash/dash/pdksh
source "/cargo/env.fish"  # For fish
/tmp/rfl-build.sh: line 16: /rfl/linux/scripts/min-tool-version.sh: No such file or directory
/tmp/rfl-build.sh: line 16: /home/user/.cargo/bin/cargo: No such file or directory
  network time: Fri, 17 May 2024 12:29:50 GMT
##[error]Process completed with exit code 127.
Post job cleanup.
```

</details>