@rust-log-analyzer 2024-05-17
The job **`rfl`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/9132168326/job/25112934162) [(plain)](https://github.com/rust-lang/rust/commit/33498b6e04fc7325afd436b4f90baeea99a882ec/checks/25112934162/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
Getting action download info
Download action repository 'msys2/setup-msys2@v2.22.0' (SHA:cc11e9188b693c2b100158c3322424c4cc1dadea)
Download action repository 'actions/checkout@v4' (SHA:0ad4b8fadaa221de15dcec353f45205ec38ea70b)
Download action repository 'actions/upload-artifact@v4' (SHA:65462800fd760344b1a7b4382951275a0abb4808)
Complete job name: PR - rfl
git config --global core.autocrlf false
shell: /usr/bin/bash --noprofile --norc -e -o pipefail {0}
---
      --llvm-root=/usr/lib/llvm-17 \
      --enable-llvm-link-shared \
      --set rust.thin-lto-import-instr-limit=10

COPY /scripts/rfl-build.sh /tmp/rfl-build.sh
ENV SCRIPT bash /tmp/rfl-build.sh

set -euo pipefail

LINUX_VERSION=v6.9
LINUX_VERSION=v6.9

# Build rustc
../x.py --stage 1 build library/std
BUILT_RUSTC=$(realpath ./build/x86_64-unknown-linux-gnu/stage1/bin/rustc)
# Install rustup so that we have cargo for installing bindgen
# Install rustup so that we have cargo for installing bindgen
curl --proto '=https' --tlsv1.2 -sSf -o rustup.sh https://sh.rustup.rs
sh rustup.sh -y --default-toolchain stable --profile minimal
source /cargo/env

mkdir -p rfl
cd rfl
cd rfl

# Download Linux
git clone --depth 1 --branch ${LINUX_VERSION} \
    https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
# Install bindgen
cargo install --locked \
cargo install --locked \
    --version $(linux/scripts/min-tool-version.sh bindgen) bindgen-cli
# Configure Linux
# Configure Linux
cat <<EOF > linux/kernel/configs/rfl-for-rust-ci.config
# CONFIG_RETHUNK is not set
# CONFIG_X86_KERNEL_IBT is not set

CONFIG_RUST=y

CONFIG_SAMPLES=y
---
CONFIG_KUNIT=y
CONFIG_RUST_KERNEL_DOCTESTS=y
EOF

make -C linux LLVM=1 RUSTC=${BUILT_RUSTC} -j$(($(nproc) + 1)) \
    rustavailable \
    defconfig \
    rfl-for-rust-ci.config

make -C linux LLVM=1 RUSTC=${BUILT_RUSTC} -j$(($(nproc) + 1)) \
    samples/rust/rust_minimal.o \
    samples/rust/rust_print.o \
    drivers/net/phy/ax88796b_rust.o

# ignore-tidy-linelength

set -ex
---
Image input checksum 53add9a3fd1f4825fa2739f2fda0eb040abb18d08d88a7512fe1a00904051f6837837553610ec1b16a0d6b233fbd2faf100e36569f6b4a2f99aa8f6b29e66264
##[group]Building docker image for rfl
Docker version 24.0.9, build 2936816
nostalgic_galois
Attempting with retry: docker buildx build --rm -t rust-ci -f /home/runner/work/rust/rust/src/ci/docker/host-x86_64/rfl/Dockerfile --cache-from type=registry,ref=ghcr.io/rust-lang-ci/rust-ci-cache:53add9a3fd1f4825fa2739f2fda0eb040abb18d08d88a7512fe1a00904051f6837837553610ec1b16a0d6b233fbd2faf100e36569f6b4a2f99aa8f6b29e66264 --output=type=docker /home/runner/work/rust/rust/src/ci/docker

#1 [internal] booting buildkit
#1 pulling image moby/buildkit:buildx-stable-1
#1 pulling image moby/buildkit:buildx-stable-1 0.2s done
---
100 22.0M  100 22.0M    0     0   146M      0 --:--:-- --:--:-- --:--:--  146M
#11 0.187 + chmod +x /usr/local/bin/sccache
#11 DONE 0.2s

#12 [5/5] COPY /scripts/rfl-build.sh /tmp/rfl-build.sh

#13 exporting to docker image format
#13 exporting layers
#13 exporting layers 47.9s done
---
DirectMap4k:      198592 kB
DirectMap2M:     9238528 kB
DirectMap1G:    26214400 kB
##[endgroup]
Executing bash /tmp/rfl-build.sh
+ bash /tmp/rfl-build.sh
    Finished `dev` profile [unoptimized] target(s) in 0.03s
##[endgroup]
##[group]Building stage0 library artifacts (x86_64-unknown-linux-gnu)
    Updating crates.io index
---
    Finished `release` profile [optimized] target(s) in 57.79s
##[endgroup]
Build completed successfully in 0:06:14
info: downloading installer
error: $HOME differs from euid-obtained home directory: you may be using sudo
error: $HOME directory: /home/user
error: euid-obtained home directory: /home/ubuntu
info: profile set to 'minimal'
info: default host triple is x86_64-unknown-linux-gnu
info: syncing channel updates for 'stable-x86_64-unknown-linux-gnu'
info: latest update on 2024-05-02, rust version 1.78.0 (9b00956e5 2024-04-29)
info: downloading component 'rust-std'
info: downloading component 'rustc'
info: installing component 'cargo'
info: installing component 'rust-std'
info: installing component 'rust-std'
info: installing component 'rustc'
info: default toolchain set to 'stable-x86_64-unknown-linux-gnu'

  stable-x86_64-unknown-linux-gnu installed - rustc 1.78.0 (9b00956e5 2024-04-29)


Rust is installed now. Great!
To get started you may need to restart your current shell.
This would reload your PATH environment variable to include
This would reload your PATH environment variable to include
Cargo's bin directory (/cargo/bin).
To configure your current shell, you need to source
the corresponding env file under /cargo.


This is usually done by running one of the following (note the leading DOT):
. "/cargo/env"            # For sh/bash/zsh/ash/dash/pdksh
source "/cargo/env.fish"  # For fish
Cloning into 'linux'...

You are in 'detached HEAD' state. You can look around, make experimental
changes and commit them, and you can discard any commits you make in this
state without impacting any branches by switching back to a branch.
---
   Compiling clap v4.2.1
   Compiling env_logger v0.10.0
   Compiling bindgen-cli v0.65.1
    Finished `release` profile [optimized] target(s) in 24.76s
  Installing /cargo/bin/bindgen
   Installed package `bindgen-cli v0.65.1` (executable `bindgen`)
make: Entering directory '/checkout/obj/rfl/linux'
***
*** Rust compiler '/checkout/obj/build/x86_64-unknown-linux-gnu/stage1/bin/rustc' is too new. This may or may not work.
***   Your version:     1.80.0
***   Expected version: 1.76.0
***
***
*** Please see Documentation/rust/quick-start.rst for details
*** on how to set up the Rust support.
Rust is available!
  HOSTCC  scripts/basic/fixdep
  HOSTCC  scripts/kconfig/conf.o
  HOSTCC  scripts/kconfig/confdata.o
---
  HOSTCC  scripts/kconfig/util.o
  HOSTCC  scripts/kconfig/lexer.lex.o
  HOSTCC  scripts/kconfig/parser.tab.o
  HOSTLD  scripts/kconfig/conf
*** Default configuration is based on 'x86_64_defconfig'
# configuration written to .config
#
Using .config as base
Using .config as base
Merging ./kernel/configs/rfl-for-rust-ci.config
Value of CONFIG_X86_KERNEL_IBT is redefined by fragment ./kernel/configs/rfl-for-rust-ci.config:
Previous value: CONFIG_X86_KERNEL_IBT=y
New value: # CONFIG_X86_KERNEL_IBT is not set

Value of CONFIG_RUST is redefined by fragment ./kernel/configs/rfl-for-rust-ci.config:
Previous value: # CONFIG_RUST is not set
New value: CONFIG_RUST=y

Value of CONFIG_SAMPLES is redefined by fragment ./kernel/configs/rfl-for-rust-ci.config:
Previous value: # CONFIG_SAMPLES is not set
New value: CONFIG_SAMPLES=y

Value of CONFIG_AX88796B_PHY is redefined by fragment ./kernel/configs/rfl-for-rust-ci.config:
Previous value: # CONFIG_AX88796B_PHY is not set
New value: CONFIG_AX88796B_PHY=y

Value of CONFIG_KUNIT is redefined by fragment ./kernel/configs/rfl-for-rust-ci.config:
Previous value: # CONFIG_KUNIT is not set
New value: CONFIG_KUNIT=y
#
# merged configuration written to .config (needs make)
#
#
---
  WRAP    arch/x86/include/generated/uapi/asm/fcntl.h
  WRAP    arch/x86/include/generated/uapi/asm/ioctl.h
  WRAP    arch/x86/include/generated/uapi/asm/ioctls.h
  UPD     include/generated/uapi/linux/version.h
  SYSHDR  arch/x86/include/generated/uapi/asm/unistd_32.h
  WRAP    arch/x86/include/generated/uapi/asm/ipcbuf.h
  SYSHDR  arch/x86/include/generated/uapi/asm/unistd_64.h
  WRAP    arch/x86/include/generated/uapi/asm/param.h
  WRAP    arch/x86/include/generated/uapi/asm/poll.h
  SYSHDR  arch/x86/include/generated/uapi/asm/unistd_x32.h
  WRAP    arch/x86/include/generated/uapi/asm/resource.h
  WRAP    arch/x86/include/generated/uapi/asm/socket.h
  WRAP    arch/x86/include/generated/uapi/asm/sockios.h
  HOSTCC  arch/x86/tools/relocs_32.o
  WRAP    arch/x86/include/generated/uapi/asm/termbits.h
  WRAP    arch/x86/include/generated/uapi/asm/termios.h
  WRAP    arch/x86/include/generated/uapi/asm/types.h
  SYSTBL  arch/x86/include/generated/asm/syscalls_32.h
  SYSHDR  arch/x86/include/generated/asm/unistd_32_ia32.h
mkdir -p /checkout/obj/rfl/linux/tools/objtool && make O=/checkout/obj/rfl/linux subdir=tools/objtool --no-print-directory -C objtool 
  WRAP    arch/x86/include/generated/asm/early_ioremap.h
  WRAP    arch/x86/include/generated/asm/mcs_spinlock.h
  WRAP    arch/x86/include/generated/asm/irq_regs.h
  WRAP    arch/x86/include/generated/asm/kmap_size.h
  HOSTCC  scripts/selinux/genheaders/genheaders
  WRAP    arch/x86/include/generated/asm/local64.h
  WRAP    arch/x86/include/generated/asm/local64.h
  WRAP    arch/x86/include/generated/asm/mmiowb.h
  WRAP    arch/x86/include/generated/asm/module.lds.h
  HOSTRUSTC scripts/generate_rust_target
  WRAP    arch/x86/include/generated/asm/rwonce.h
  HOSTCC  scripts/selinux/mdp/mdp
  WRAP    arch/x86/include/generated/asm/unaligned.h
  HOSTCC  scripts/kallsyms
  SYSHDR  arch/x86/include/generated/asm/unistd_64_x32.h
  HOSTCC  /checkout/obj/rfl/linux/tools/objtool/fixdep.o
  SYSTBL  arch/x86/include/generated/asm/syscalls_64.h
  HOSTCC  arch/x86/tools/relocs_64.o
  HOSTLD  /checkout/obj/rfl/linux/tools/objtool/fixdep-in.o
  HOSTCC  scripts/sorttable
  LINK    /checkout/obj/rfl/linux/tools/objtool/fixdep
  HOSTCC  scripts/asn1_compiler
  HOSTRUSTC scripts/rustdoc_test_builder
  HOSTRUSTC scripts/rustdoc_test_gen
  HOSTRUSTC scripts/rustdoc_test_gen
  CC      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/exec-cmd.o
  CC      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/help.o
  HOSTLD  arch/x86/tools/relocs
  CC      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/pager.o
  CC      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/parse-options.o
  CC      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/run-command.o
  CC      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/sigchain.o
  INSTALL /checkout/obj/rfl/linux/tools/objtool/libsubcmd/include/subcmd/exec-cmd.h
  INSTALL /checkout/obj/rfl/linux/tools/objtool/libsubcmd/include/subcmd/help.h
  UPD     scripts/target.json
  INSTALL /checkout/obj/rfl/linux/tools/objtool/libsubcmd/include/subcmd/pager.h
  INSTALL /checkout/obj/rfl/linux/tools/objtool/libsubcmd/include/subcmd/parse-options.h
  INSTALL /checkout/obj/rfl/linux/tools/objtool/libsubcmd/include/subcmd/run-command.h
  CC      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/subcmd-config.o
  INSTALL libsubcmd_headers
  UPD     include/generated/utsrelease.h
  UPD     include/generated/utsrelease.h
  LD      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/libsubcmd-in.o
  AR      /checkout/obj/rfl/linux/tools/objtool/libsubcmd/libsubcmd.a
/bin/sh: 1: llvm-ar: not found
make[4]: *** [Makefile:83: /checkout/obj/rfl/linux/tools/objtool/libsubcmd/libsubcmd.a] Error 127
make[3]: *** [Makefile:82: /checkout/obj/rfl/linux/tools/objtool/libsubcmd/libsubcmd.a] Error 2
make[2]: *** [Makefile:72: objtool] Error 2
make[1]: *** [/checkout/obj/rfl/linux/Makefile:1358: tools/objtool] Error 2
make[1]: *** Waiting for unfinished jobs....
  HOSTCC  scripts/mod/mk_elfconfig
  CC      scripts/mod/devicetable-offsets.s
  UPD     scripts/mod/devicetable-offsets.h
  MKELF   scripts/mod/elfconfig.h
  MKELF   scripts/mod/elfconfig.h
  HOSTCC  scripts/mod/modpost.o
  HOSTCC  scripts/mod/file2alias.o
  HOSTCC  scripts/mod/sumversion.o
  HOSTCC  scripts/mod/symsearch.o
  HOSTLD  scripts/mod/modpost
  CC      kernel/bounds.s
  CHKSHA1 include/linux/atomic/atomic-arch-fallback.h
  CHKSHA1 include/linux/atomic/atomic-instrumented.h
  CHKSHA1 include/linux/atomic/atomic-long.h
  UPD     include/generated/bounds.h
  CC      arch/x86/kernel/asm-offsets.s
  UPD     include/generated/asm-offsets.h
  CALL    scripts/checksyscalls.sh
  CALL    scripts/checksyscalls.sh
make: Leaving directory '/checkout/obj/rfl/linux'
make: *** [Makefile:240: __sub-make] Error 2
  network time: Fri, 17 May 2024 17:54:18 GMT
##[error]Process completed with exit code 2.
Post job cleanup.
```

</details>